+++ src/App.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="container">
     <div>
-      <h1>Routing</h1>
+      <h1>Critical Asset Management System</h1>
       <hr />
       <nav-bar></nav-bar>
       <router-view></router-view>
