+++ src/components/NavBar.vue
@@ -2,7 +2,6 @@
   <div>
     <b-nav pills>
       <b-nav-item to="/" active-class="active" exact>Home</b-nav-item>
-      <b-nav-item to="/about" active-class="active">About</b-nav-item>
       <b-nav-item to="/userform" active-class="active">User Form</b-nav-item>
       <b-nav-item to="/areaform" active-class="active">Area Form</b-nav-item>
       <b-nav-item to="/assetform" active-class="active">Asset Form</b-nav-item>
